requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
groq>=0.4.2
python-dotenv>=1.0.1
pydantic>=2.5.0
